Make Player able to shoot. Heap management is bad, fix it in next commit
Bunch of errors with memory exist when trying to implement bullets, heap
management is poor due to incorrect use of delete and new and pointer
management. Fix this by using vectors and reducing use of raw pointers.
Still need to delete SDL_Textures after use, this is only commented out
for now until it is properly implemented. (Copy constructor that is
called with vector.push_back() method is called destructor on a texture
that is still in use and has just been pushed to a vector, resulting in
memory issues, but also SDL_Textures can't be copied so tough to create
a copy constructor, solve this issue for the next commit so that heap
management is solved.
Fixed projectile drawing, previously the projectiles were drawn before
the map and the player so they were unable to be seen but were there