Fix memory issues that exist when destructing SDL_Textures after use
When destructing SDL_Textures in ObjTexture class destructor, tons of
memory errors would occur. Change from last commit to fix this is for
objects that have an ObjTexture member variable, to have a pointer to it
instead. This way, when pushing that ObjTexture into a vector, instead
of the vector copying the ObjTexture (which has no copy constructor and
ends up shallow copying the SDL_Texture pointer), it copies the pointer
to the ObjTexture and puts it in the vector. The reason this is better
is that if instead of an ObjTexture pointer, an ObjTexture was passed
into the vector, then when the object is copied into the vector and
destroyed, both the copied object and the new object's SDL_Texture
pointer points to the same memory and the one in the vector no longer
has a valid SDL_Texture. This is the problem that arises when you use a
memory address stored in a local variable and then push it to a global
variable (Projectile vector); when the local variable gets destroyed it
messes with the global vector's memory (and that's why pointers help)