Put main game loop in a separate function that is called within main
In Main, call function startWave() that starts the game and returns when
game is over. Then return value is checked to see if Survivor is dead,
and if so, show the game over screen. If Survivor is not dead, then when
multiple waves are implemented, the next wave will start