Create ZombieManager class
Instead of having Vector of Zombies as a global variable, and then
iterate through it in the main game loop whenever updating, rendering,
or doing anything with the zombies, create a ZombieManager class to do
all of these things. Cleaner, more separated code and more Object
Oriented. Potentially need to fix a couple things, such as collision
detection (still iterates through Zombies in main loop instead of in a
ZombieManager method) and resetting for the next wave or game