Implement Wave Complete screen
Add "Buy Health" and "Buy Ammo" buttons for the cost of score, and add
"Continue" button to continue to the next wave. Also use Container class
to make writing to screen way cleaner and simpler, and do this to
simplify the gameOver and printInfo functions.

Make the buttons in the Wave Complete screen change color when the mouse
hovers over them to show that it's being selected

Make ammo more scarce since it can be purchased now (still need to add
Survivor combat attacks for when you run out of ammo)